requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Werkzeug==2.3.7
mysql-connector-python==8.1.0
python-dotenv==1.0.0
boto3==1.34.69
botocore==1.34.69
click==8.1.8
gunicorn==21.2.0